requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
rembg[cpu]>=2.0.50
scikit-learn>=1.3.0
opensimplex>=0.4.5
python-multipart>=0.0.6
